Move the inserter tests out of the gallery suite
The tests cover the inserter next to the parent selector, not the
gallery, so they get their own file. The gallery test now uses a data
URI image, since the block structure it asserts does not need the
media library.
