+++ test/e2e/specs/editor/blocks/gallery.spec.js
@@ -29,94 +29,6 @@ test.describe( 'Gallery', () => {
 		await requestUtils.deleteAllMedia();
 	} );
 
-	test( 'adds an image after the selected one from the parent selector', async ( {
-		admin,
-		editor,
-		page,
-	} ) => {
-		await admin.createNewPost();
-		await editor.insertBlock( {
-			name: 'core/gallery',
-			innerBlocks: [
-				{
-					name: 'core/image',
-					attributes: { url: uploadedMedia.source_url, alt: 'First' },
-				},
-				{
-					name: 'core/image',
-					attributes: {
-						url: uploadedMedia.source_url,
-						alt: 'Second',
-					},
-				},
-			],
-		} );
-		await editor.canvas
-			.locator( '[data-type="core/image"]' )
-			.first()
-			.click();
-
-		await editor.showBlockToolbar();
-		await page.locator( 'role=button[name="Add Image"]' ).click();
-
-		// The new empty image lands between the two, selected, showing its
-		// media placeholder.
-		await expect.poll( editor.getBlocks ).toMatchObject( [
-			{
-				name: 'core/gallery',
-				innerBlocks: [
-					{ name: 'core/image', attributes: { alt: 'First' } },
-					{ name: 'core/image', attributes: { alt: '' } },
-					{ name: 'core/image', attributes: { alt: 'Second' } },
-				],
-			},
-		] );
-	} );
-
-	test( 'hides the parent selector inserter only inside text flow wrappers', async ( {
-		admin,
-		editor,
-		page,
-	} ) => {
-		await admin.createNewPost();
-		await editor.insertBlock( {
-			name: 'core/group',
-			attributes: { layout: { type: 'constrained' } },
-			innerBlocks: [
-				{ name: 'core/paragraph', attributes: { content: 'Text' } },
-			],
-		} );
-		await editor.canvas.locator( '[data-type="core/paragraph"]' ).click();
-
-		await editor.showBlockToolbar();
-		const toolbar = page.locator( 'role=toolbar[name="Block tools"i]' );
-		await expect(
-			toolbar.locator( 'role=button[name="Select parent block: Group"]' )
-		).toBeVisible();
-		await expect(
-			toolbar.locator( 'role=button[name="Add block"]' )
-		).toBeVisible();
-
-		// A list merges with the text flow: Enter continues it, so its
-		// items get no inserter.
-		await admin.createNewPost();
-		await editor.insertBlock( {
-			name: 'core/list',
-			innerBlocks: [
-				{ name: 'core/list-item', attributes: { content: 'one' } },
-			],
-		} );
-		await editor.canvas.locator( '[data-type="core/list-item"]' ).click();
-
-		await editor.showBlockToolbar();
-		await expect(
-			toolbar.locator( 'role=button[name="Select parent block: List"]' )
-		).toBeVisible();
-		await expect(
-			toolbar.locator( 'role=button[name^="Add "]' )
-		).toBeHidden();
-	} );
-
 	test( 'can be transformed from pasting shortcode, and can undo/redo', async ( {
 		admin,
 		editor,
